create type declaration file for nacl
