+++ packages/lisk-cryptography/src/nacl/index.ts
@@ -12,30 +12,8 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-import { KeypairBytes } from '../keys';
+import { NaclInterface } from '../types/nacl';
 
-export interface NaclInterface {
-	box(
-		messageInBytes: Buffer,
-		nonceInBytes: Buffer,
-		convertedPublicKey: Buffer,
-		convertedPrivateKey: Buffer,
-	): Buffer;
-	getKeyPair(hashedSeed: Buffer): KeypairBytes;
-	getRandomBytes(length: number): Buffer;
-	openBox(
-		cipherBytes: Buffer,
-		nonceBytes: Buffer,
-		convertedPublicKey: Buffer,
-		convertedPrivateKey: Buffer,
-	): Buffer;
-	signDetached(messageBytes: Buffer, privateKeyBytes: Buffer): Buffer;
-	verifyDetached(
-		messageBytes: Buffer,
-		signatureBytes: Buffer,
-		publicKeyBytes: Buffer,
-	): boolean;
-}
 // tslint:disable-next-line no-let
 let lib: NaclInterface;
